+++ packages/sui-js/test/stringSpec.js
@@ -3,14 +3,15 @@ import {expect} from 'chai'
 import {parseQueryString, toQueryString} from '../src/string/index'
 
 describe('@s-ui/js', () => {
-  describe('when working with query params', () => {
-    it('should convert to string', () => {
+  describe('string:toQueryString', () => {
+    it('should convert object params to query string', () => {
       const params = {a: 1, b: 'test'}
       const query = 'a=1&b=test'
       expect(toQueryString(params)).to.be.equal(query)
     })
-
-    it('should parse to object', () => {
+  })
+  describe('string:parseQueryString', () => {
+    it('should convert query string to object params', () => {
       const query = '?a=1&b=test'
       const params = {a: '1', b: 'test'}
       expect(parseQueryString(query)).to.deep.equal(params)